+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2021-02-23
+
+### Added
+- Billing options
+- App Store metadata
+
 ## [1.8.0] - 2021-01-18
 
 ### Fixed